requirements.txt
# Core data and computation
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
alpha-vantage>=2.3.1
ccxt>=4.0.0
requests>=2.31.0
urllib3>=1.26.0
certifi>=2023.7.22
requests-cache>=1.1.0

# Visualization and UI
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Technical analysis
ta>=0.10.2
TA-Lib>=0.4.26

# Financial calculations
quantlib>=1.31
scipy>=1.11.0
scikit-learn>=1.3.0

# Data validation and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
cachetools>=5.3.0

# API clients
newsapi-python>=0.2.6
tweepy>=4.14.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0
coverage>=7.0.0
